DEV: Fix admin route checker import
This cannot be in models/admin-plugin because other
parts of the app need it too
