+++ app/assets/javascripts/admin/addon/controllers/admin-plugins-index.js
@@ -1,8 +1,8 @@
 import Controller from "@ember/controller";
 import { action } from "@ember/object";
 import { service } from "@ember/service";
+import { adminRouteValid } from "discourse/lib/admin-utilities";
 import { popupAjaxError } from "discourse/lib/ajax-error";
-import { adminRouteValid } from "admin/models/admin-plugin";
 import SiteSetting from "admin/models/site-setting";
 
 export default class AdminPluginsIndexController extends Controller {
